requirements.txt
gensim==4.3.2
joblib==1.4.2
keras==3.4.1
matplotlib==3.8.4
numpy==2.0.0
pandas==2.2.2
Pillow==10.3.0
plotly==5.22.0
pyparsing==3.1.2
python-dotenv==1.0.1
scikit_learn==1.5.0
scikit_surprise==1.1.4
scipy==1.14.0
seaborn==0.13.2
streamlit==1.36.0
tensorflow==2.16.1
